piopio/requirements.txt
Django==2.2.5
django-cors-headers
djangorestframework
Pillow>=6.1.0
mysqlclient
djangorestframework_simplejwt